@@ -6,12 +6,14 @@ package cmd
 import (
 	"context"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"path"
 	"path/filepath"
 	"strings"
 	"time"
+
+	"golang.org/x/text/cases"
+	"golang.org/x/text/language"
 
 	ops "github.com/Azure/aks-engine-azurestack/cmd/rotatecerts"
 	"github.com/Azure/aks-engine-azurestack/pkg/api"
@@ -171,7 +173,7 @@ func (rcc *rotateCertsCmd) validateArgs() (err error) {
 			return errors.Errorf("error creating output directory (%s)", rcc.outputDirectory)
 		}
 	}
-	if _, err := ioutil.ReadDir(rcc.outputDirectory); err != nil {
+	if _, err := os.ReadDir(rcc.outputDirectory); err != nil {
 		return errors.Wrapf(err, "reading output directory %s", rcc.outputDirectory)
 	}
 	return nil
@@ -346,7 +348,8 @@ func (rcc *rotateCertsCmd) getAgentNodes() (nodeMap, error) {
 			Port:    22,
 			Jumpbox: rcc.jumpbox,
 		}
-		switch api.OSType(strings.Title(nli.Status.NodeInfo.OperatingSystem)) {
+		caser := cases.Title(language.English)
+		switch api.OSType(caser.String(nli.Status.NodeInfo.OperatingSystem)) {
 		case api.Linux:
 			node.OperatingSystem = api.Linux
 			node.AuthConfig = rcc.linuxAuthConfig
@@ -650,7 +653,7 @@ func (rcc *rotateCertsCmd) getKubeClient() (*kubernetes.CompositeClientSet, erro
 	configPathSuffix := path.Join("kubeconfig", fmt.Sprintf("kubeconfig.%s.json", rcc.location))
 
 	oldConfigPath := path.Join(rcc.backupDirectory, configPathSuffix)
-	oldConfig, err := ioutil.ReadFile(oldConfigPath)
+	oldConfig, err := os.ReadFile(oldConfigPath)
 	if err != nil {
 		return nil, errors.Wrapf(err, "reading %s", oldConfigPath)
 	}
@@ -660,7 +663,7 @@ func (rcc *rotateCertsCmd) getKubeClient() (*kubernetes.CompositeClientSet, erro
 	}
 
 	newConfigPath := path.Join(rcc.outputDirectory, configPathSuffix)
-	newConfig, err := ioutil.ReadFile(newConfigPath)
+	newConfig, err := os.ReadFile(newConfigPath)
 	if err != nil {
 		return nil, errors.Wrapf(err, "reading %s", newConfigPath)
 	}
