@@ -5,8 +5,8 @@ package azurestack
 
 import (
 	"fmt"
-	"io/ioutil"
 	"net/http"
+	"os"
 
 	"github.com/Azure/aks-engine-azurestack/pkg/armhelpers/azurestack/testserver"
 	"github.com/Azure/azure-sdk-for-go/services/compute/mgmt/2017-03-30/compute"
@@ -615,7 +615,7 @@ func (mc *HTTPMockClient) RegisterVMImageFetcherInterface() {
 }
 
 func readFromFile(filePath string) (string, error) {
-	bytes, err := ioutil.ReadFile(filePath)
+	bytes, err := os.ReadFile(filePath)
 	if err != nil {
 		return "", fmt.Errorf("Fail to read file %q , err -  %q", filePath, err)
 	}
