@@ -10,7 +10,6 @@ import (
 	"bufio"
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
 	"log"
 	"math/rand"
 	"os"
@@ -66,6 +65,7 @@ type Config struct {
 	RunVMSSNodePrototype               bool   `envconfig:"RUN_VMSS_NODE_PROTOTYPE" default:"false"`
 	AddNodePoolInput                   string `envconfig:"ADD_NODE_POOL_INPUT" default:""`
 	TestPVC                            bool   `envconfig:"TEST_PVC" default:"false"`
+	CleanPVC                           bool   `envconfig:"CLEAN_PVC" default:"true"`
 	SubscriptionID                     string `envconfig:"SUBSCRIPTION_ID"`
 	ClientID                           string `envconfig:"CLIENT_ID"`
 	ClientSecret                       string `envconfig:"CLIENT_SECRET"`
@@ -170,7 +170,7 @@ func (c *Config) UpdateCustomCloudClusterDefinition(ccc *CustomCloudConfig) erro
 	if err != nil {
 		return fmt.Errorf("Error fail to marshal containerService object %p", err)
 	}
-	err = ioutil.WriteFile(clusterDefinitionFullPath, csBytes, 644)
+	err = os.WriteFile(clusterDefinitionFullPath, csBytes, 644)
 	if err != nil {
 		return fmt.Errorf("Error fail to write file object %p", err)
 	}
@@ -179,7 +179,7 @@ func (c *Config) UpdateCustomCloudClusterDefinition(ccc *CustomCloudConfig) erro
 
 func parseVlabsContainerSerice(clusterDefinitionFullPath string) api.VlabsARMContainerService {
 
-	bytes, err := ioutil.ReadFile(clusterDefinitionFullPath)
+	bytes, err := os.ReadFile(clusterDefinitionFullPath)
 	if err != nil {
 		log.Fatalf("Error while trying to read cluster definition at (%s):%s\n", clusterDefinitionFullPath, err)
 	}
@@ -310,7 +310,7 @@ func (c *Config) SetEnvVars() error {
 // ReadPublicSSHKey will read the contents of the public ssh key on disk into a string
 func (c *Config) ReadPublicSSHKey() (string, error) {
 	file := c.GetSSHKeyPath() + ".pub"
-	contents, err := ioutil.ReadFile(file)
+	contents, err := os.ReadFile(file)
 	if err != nil {
 		log.Printf("Error while trying to read public ssh key at (%s):%s\n", file, err)
 		return "", err
