@@ -57,13 +57,15 @@ const (
 
 const (
 	// KubernetesDefaultRelease is the default Kubernetes release
-	KubernetesDefaultRelease string = "1.23"
+	KubernetesDefaultRelease string = "1.24"
 	// KubernetesDefaultReleaseWindows is the default Kubernetes release for Windows
-	KubernetesDefaultReleaseWindows string = "1.23"
+	KubernetesDefaultReleaseWindows string = "1.24"
 	// KubernetesDefaultReleaseAzureStack is the default Kubernetes release on Azure Stack
-	KubernetesDefaultReleaseAzureStack string = "1.23"
+	KubernetesDefaultReleaseAzureStack string = "1.24"
 	// KubernetesDefaultReleaseWindowsAzureStack is the default Kubernetes release for Windows on Azure Stack
-	KubernetesDefaultReleaseWindowsAzureStack string = "1.23"
+	KubernetesDefaultReleaseWindowsAzureStack string = "1.24"
+	// PodSecurityPolicyRemovedVersion is the first Kubernetes version that does not includes the PSP admission plugin
+	PodSecurityPolicyRemovedVersion = "1.25.0"
 )
 
 const LegacyControlPlaneVMPrefix string = "k8s-master"
