@@ -340,7 +340,7 @@ var AllKubernetesSupportedVersions = map[string]bool{
 	"1.23.12":        false,
 	"1.23.13":        false,
 	"1.23.14":        false,
-	"1.23.15":        true,
+	"1.23.15":        false,
 	"1.24.0-alpha.2": false,
 	"1.24.0-alpha.3": false,
 	"1.24.0":         false,
@@ -353,6 +353,7 @@ var AllKubernetesSupportedVersions = map[string]bool{
 	"1.24.7":         false,
 	"1.24.8":         false,
 	"1.24.9":         true,
+	"1.25.7":         true,
 }
 
 // AllKubernetesSupportedVersionsAzureStack is a hash table of all supported Kubernetes version strings on Azure Stack
@@ -393,9 +394,10 @@ var AllKubernetesSupportedVersionsAzureStack = map[string]bool{
 	"1.23.12": false,
 	"1.23.13": false,
 	"1.23.14": false,
-	"1.23.15": true,
+	"1.23.15": false,
 	"1.24.7":  false,
 	"1.24.9":  true,
+	"1.25.7":  true,
 }
 
 // AllKubernetesWindowsSupportedVersionsAzureStack maintain a set of available k8s Windows versions in aks-engine on Azure Stack
@@ -431,9 +433,10 @@ var AllKubernetesWindowsSupportedVersionsAzureStack = map[string]bool{
 	"1.23.12": false,
 	"1.23.13": false,
 	"1.23.14": false,
-	"1.23.15": true,
+	"1.23.15": false,
 	"1.24.7":  false,
 	"1.24.9":  true,
+	"1.25.7":  true,
 }
 
 // GetDefaultKubernetesVersion returns the default Kubernetes version, that is the latest patch of the default release
