@@ -38,7 +38,7 @@ var (
 		"3.1.0", "3.1.1", "3.1.2", "3.1.2", "3.1.3", "3.1.4", "3.1.5", "3.1.6", "3.1.7", "3.1.8", "3.1.9", "3.1.10",
 		"3.2.0", "3.2.1", "3.2.2", "3.2.3", "3.2.4", "3.2.5", "3.2.6", "3.2.7", "3.2.8", "3.2.9", "3.2.11", "3.2.12",
 		"3.2.13", "3.2.14", "3.2.15", "3.2.16", "3.2.23", "3.2.24", "3.2.25", "3.2.26", "3.3.0", "3.3.1", "3.3.8", "3.3.9", "3.3.10", "3.3.13", "3.3.15", "3.3.18", "3.3.19", "3.3.22", "3.3.25"}
-	containerdValidVersions              = [...]string{"1.3.2", "1.3.3", "1.3.4", "1.3.5", "1.3.6", "1.3.7", "1.3.8", "1.3.9", "1.4.4", "1.4.6", "1.4.7", "1.4.8", "1.4.9", "1.4.11", "1.5.11", "1.5.13"}
+	containerdValidVersions              = [...]string{"1.3.2", "1.3.3", "1.3.4", "1.3.5", "1.3.6", "1.3.7", "1.3.8", "1.3.9", "1.4.4", "1.4.6", "1.4.7", "1.4.8", "1.4.9", "1.4.11", "1.5.11", "1.5.13", "1.5.16"}
 	kubernetesImageBaseTypeValidVersions = [...]string{"", common.KubernetesImageBaseTypeGCR, common.KubernetesImageBaseTypeMCR}
 	cachingTypesValidValues              = [...]string{"", string(compute.CachingTypesNone), string(compute.CachingTypesReadWrite), string(compute.CachingTypesReadOnly)}
 	linuxEth0MTUAllowedValues            = [...]int{1500, 3900}
@@ -418,6 +418,10 @@ func (a *Properties) validateMasterProfile(isUpdate bool) error {
 		if m.Distro != "" && !m.IsUbuntu() {
 			return errors.Errorf("auditd was enabled for master vms, but an Ubuntu-based distro was not selected")
 		}
+	} else {
+		if a.FeatureFlags.IsEnforceUbuntu2004DisaStigEnabled() && m.Distro != "" && m.IsUbuntu() {
+			return errors.New("AuditD should be enabled in all Ubuntu-based pools if feature flag 'EnforceUbuntu2004DisaStig' is set")
+		}
 	}
 
 	var validOSDiskCachingType bool
@@ -483,6 +487,10 @@ func (a *Properties) validateAgentPoolProfiles(isUpdate bool) error {
 		if to.Bool(agentPoolProfile.AuditDEnabled) {
 			if agentPoolProfile.Distro != "" && !agentPoolProfile.IsUbuntu() {
 				return errors.Errorf("You have enabled auditd in agent pool %s, but you did not specify an Ubuntu-based distro", agentPoolProfile.Name)
+			}
+		} else {
+			if a.FeatureFlags.IsEnforceUbuntu2004DisaStigEnabled() && agentPoolProfile.IsUbuntu() {
+				return errors.New("AuditD should be enabled in all Ubuntu-based pools if feature flag 'EnforceUbuntu2004DisaStig' is set")
 			}
 		}
 
@@ -789,6 +797,12 @@ func (a *Properties) validateAddons(isUpdate bool) error {
 				case common.SecretsStoreCSIDriverAddonName:
 					if !common.IsKubernetesVersionGe(a.OrchestratorProfile.OrchestratorVersion, "1.16.0") {
 						return errors.Errorf("%s add-on can only be used in 1.16+", addon.Name)
+					}
+				case common.PodSecurityPolicyAddonName:
+					if common.ShouldDisablePodSecurityPolicyAddon(a.OrchestratorProfile.OrchestratorVersion) {
+						log.Warn("The PodSecurityPolicy admission was removed in Kubernetes v1.25+. " +
+							"The pod security standards will be enforced by the built-in PodSecurity admission controller instead. " +
+							"See https://github.com/Azure/aks-engine-azurestack/blob/master/docs/topics/pod-security.md")
 					}
 				case common.AzureArcOnboardingAddonName:
 					if err := addon.validateArcAddonConfig(); err != nil {
