@@ -147,20 +147,21 @@ func TestK8sVars(t *testing.T) {
 		"primaryAvailabilitySetName":         "",
 		"primaryScaleSetName":                cs.Properties.GetPrimaryScaleSetName(),
 		"cloudInitFiles": map[string]interface{}{
-			"provisionScript":                getBase64EncodedGzippedCustomScript(kubernetesCSEMainScript, cs),
-			"provisionSource":                getBase64EncodedGzippedCustomScript(kubernetesCSEHelpersScript, cs),
-			"provisionInstalls":              getBase64EncodedGzippedCustomScript(kubernetesCSEInstall, cs),
-			"provisionConfigs":               getBase64EncodedGzippedCustomScript(kubernetesCSEConfig, cs),
-			"customSearchDomainsScript":      getBase64EncodedGzippedCustomScript(kubernetesCustomSearchDomainsScript, cs),
-			"etcdSystemdService":             getBase64EncodedGzippedCustomScript(etcdSystemdService, cs),
-			"dhcpv6ConfigurationScript":      getBase64EncodedGzippedCustomScript(dhcpv6ConfigurationScript, cs),
-			"dhcpv6SystemdService":           getBase64EncodedGzippedCustomScript(dhcpv6SystemdService, cs),
-			"kubeletSystemdService":          getBase64EncodedGzippedCustomScript(kubeletSystemdService, cs),
-			"etcdMonitorSystemdService":      getBase64EncodedGzippedCustomScript(etcdMonitorSystemdService, cs),
-			"healthMonitorScript":            getBase64EncodedGzippedCustomScript(kubernetesHealthMonitorScript, cs),
-			"kubeletMonitorSystemdService":   getBase64EncodedGzippedCustomScript(kubernetesKubeletMonitorSystemdService, cs),
-			"apiserverMonitorSystemdService": getBase64EncodedGzippedCustomScript(apiserverMonitorSystemdService, cs),
-			"dockerMonitorSystemdService":    getBase64EncodedGzippedCustomScript(kubernetesDockerMonitorSystemdService, cs),
+			"provisionScript":                 getBase64EncodedGzippedCustomScript(kubernetesCSEMainScript, cs),
+			"provisionSource":                 getBase64EncodedGzippedCustomScript(kubernetesCSEHelpersScript, cs),
+			"provisionInstalls":               getBase64EncodedGzippedCustomScript(kubernetesCSEInstall, cs),
+			"provisionConfigs":                getBase64EncodedGzippedCustomScript(kubernetesCSEConfig, cs),
+			"customSearchDomainsScript":       getBase64EncodedGzippedCustomScript(kubernetesCustomSearchDomainsScript, cs),
+			"etcdSystemdService":              getBase64EncodedGzippedCustomScript(etcdSystemdService, cs),
+			"dhcpv6ConfigurationScript":       getBase64EncodedGzippedCustomScript(dhcpv6ConfigurationScript, cs),
+			"dhcpv6SystemdService":            getBase64EncodedGzippedCustomScript(dhcpv6SystemdService, cs),
+			"kubeletSystemdService":           getBase64EncodedGzippedCustomScript(kubeletSystemdService, cs),
+			"etcdMonitorSystemdService":       getBase64EncodedGzippedCustomScript(etcdMonitorSystemdService, cs),
+			"healthMonitorScript":             getBase64EncodedGzippedCustomScript(kubernetesHealthMonitorScript, cs),
+			"kubeletMonitorSystemdService":    getBase64EncodedGzippedCustomScript(kubernetesKubeletMonitorSystemdService, cs),
+			"apiServerAdmissionConfiguration": getBase64EncodedGzippedCustomScript(apiServerAdmissionConfiguration, cs),
+			"apiserverMonitorSystemdService":  getBase64EncodedGzippedCustomScript(apiserverMonitorSystemdService, cs),
+			"dockerMonitorSystemdService":     getBase64EncodedGzippedCustomScript(kubernetesDockerMonitorSystemdService, cs),
 		},
 		"provisionScriptParametersCommon":           "[concat('" + cs.GetProvisionScriptParametersCommon(api.ProvisionScriptParametersInput{Location: common.WrapAsARMVariable("location"), ResourceGroup: common.WrapAsARMVariable("resourceGroup"), TenantID: common.WrapAsARMVariable("tenantID"), SubscriptionID: common.WrapAsARMVariable("subscriptionId"), ClientID: common.WrapAsARMVariable("servicePrincipalClientId"), ClientSecret: common.WrapAsARMVariable("singleQuote") + common.WrapAsARMVariable("servicePrincipalClientSecret") + common.WrapAsARMVariable("singleQuote"), APIServerCertificate: common.WrapAsParameter("apiServerCertificate"), KubeletPrivateKey: common.WrapAsParameter("clientPrivateKey"), ClusterKeyVaultName: common.WrapAsARMVariable("clusterKeyVaultName")}) + "')]",
 		"provisionScriptParametersMaster":           "[concat('COSMOS_URI= MASTER_VM_NAME=',variables('masterVMNames')[variables('masterOffset')],' ETCD_PEER_URL=',variables('masterEtcdPeerURLs')[variables('masterOffset')],' ETCD_CLIENT_URL=',variables('masterEtcdClientURLs')[variables('masterOffset')],' MASTER_NODE=true NO_OUTBOUND=false AUDITD_ENABLED=false CLUSTER_AUTOSCALER_ADDON=false APISERVER_PRIVATE_KEY=',parameters('apiServerPrivateKey'),' CA_CERTIFICATE=',parameters('caCertificate'),' CA_PRIVATE_KEY=',parameters('caPrivateKey'),' MASTER_FQDN=',variables('masterFqdnPrefix'),' KUBECONFIG_CERTIFICATE=',parameters('kubeConfigCertificate'),' KUBECONFIG_KEY=',parameters('kubeConfigPrivateKey'),' ETCD_SERVER_CERTIFICATE=',parameters('etcdServerCertificate'),' ETCD_CLIENT_CERTIFICATE=',parameters('etcdClientCertificate'),' ETCD_SERVER_PRIVATE_KEY=',parameters('etcdServerPrivateKey'),' ETCD_CLIENT_PRIVATE_KEY=',parameters('etcdClientPrivateKey'),' ETCD_PEER_CERTIFICATES=',string(variables('etcdPeerCertificates')),' ETCD_PEER_PRIVATE_KEYS=',string(variables('etcdPeerPrivateKeys')),' ENABLE_AGGREGATED_APIS=',string(parameters('enableAggregatedAPIs')),' KUBECONFIG_SERVER=',variables('kubeconfigServer'))]",
@@ -223,22 +224,23 @@ func TestK8sVars(t *testing.T) {
 		t.Fatal(err)
 	}
 	expectedMap["cloudInitFiles"] = map[string]interface{}{
-		"provisionScript":                getBase64EncodedGzippedCustomScript(kubernetesCSEMainScript, cs),
-		"provisionSource":                getBase64EncodedGzippedCustomScript(kubernetesCSEHelpersScript, cs),
-		"provisionInstalls":              getBase64EncodedGzippedCustomScript(kubernetesCSEInstall, cs),
-		"provisionConfigs":               getBase64EncodedGzippedCustomScript(kubernetesCSEConfig, cs),
-		"customSearchDomainsScript":      getBase64EncodedGzippedCustomScript(kubernetesCustomSearchDomainsScript, cs),
-		"etcdSystemdService":             getBase64EncodedGzippedCustomScript(etcdSystemdService, cs),
-		"dhcpv6ConfigurationScript":      getBase64EncodedGzippedCustomScript(dhcpv6ConfigurationScript, cs),
-		"dhcpv6SystemdService":           getBase64EncodedGzippedCustomScript(dhcpv6SystemdService, cs),
-		"kubeletSystemdService":          getBase64EncodedGzippedCustomScript(kubeletSystemdService, cs),
-		"etcdMonitorSystemdService":      getBase64EncodedGzippedCustomScript(etcdMonitorSystemdService, cs),
-		"healthMonitorScript":            getBase64EncodedGzippedCustomScript(kubernetesHealthMonitorScript, cs),
-		"kubeletMonitorSystemdService":   getBase64EncodedGzippedCustomScript(kubernetesKubeletMonitorSystemdService, cs),
-		"apiserverMonitorSystemdService": getBase64EncodedGzippedCustomScript(apiserverMonitorSystemdService, cs),
-		"dockerMonitorSystemdService":    getBase64EncodedGzippedCustomScript(kubernetesDockerMonitorSystemdService, cs),
-		"untaintNodesScript":             getBase64EncodedGzippedCustomScript(untaintNodesScript, cs),
-		"untaintNodesSystemdService":     getBase64EncodedGzippedCustomScript(untaintNodesSystemdService, cs),
+		"provisionScript":                 getBase64EncodedGzippedCustomScript(kubernetesCSEMainScript, cs),
+		"provisionSource":                 getBase64EncodedGzippedCustomScript(kubernetesCSEHelpersScript, cs),
+		"provisionInstalls":               getBase64EncodedGzippedCustomScript(kubernetesCSEInstall, cs),
+		"provisionConfigs":                getBase64EncodedGzippedCustomScript(kubernetesCSEConfig, cs),
+		"customSearchDomainsScript":       getBase64EncodedGzippedCustomScript(kubernetesCustomSearchDomainsScript, cs),
+		"etcdSystemdService":              getBase64EncodedGzippedCustomScript(etcdSystemdService, cs),
+		"dhcpv6ConfigurationScript":       getBase64EncodedGzippedCustomScript(dhcpv6ConfigurationScript, cs),
+		"dhcpv6SystemdService":            getBase64EncodedGzippedCustomScript(dhcpv6SystemdService, cs),
+		"kubeletSystemdService":           getBase64EncodedGzippedCustomScript(kubeletSystemdService, cs),
+		"etcdMonitorSystemdService":       getBase64EncodedGzippedCustomScript(etcdMonitorSystemdService, cs),
+		"healthMonitorScript":             getBase64EncodedGzippedCustomScript(kubernetesHealthMonitorScript, cs),
+		"kubeletMonitorSystemdService":    getBase64EncodedGzippedCustomScript(kubernetesKubeletMonitorSystemdService, cs),
+		"apiServerAdmissionConfiguration": getBase64EncodedGzippedCustomScript(apiServerAdmissionConfiguration, cs),
+		"apiserverMonitorSystemdService":  getBase64EncodedGzippedCustomScript(apiserverMonitorSystemdService, cs),
+		"dockerMonitorSystemdService":     getBase64EncodedGzippedCustomScript(kubernetesDockerMonitorSystemdService, cs),
+		"untaintNodesScript":              getBase64EncodedGzippedCustomScript(untaintNodesScript, cs),
+		"untaintNodesSystemdService":      getBase64EncodedGzippedCustomScript(untaintNodesSystemdService, cs),
 	}
 
 	diff = cmp.Diff(varMap, expectedMap)
@@ -279,6 +281,7 @@ func TestK8sVars(t *testing.T) {
 		"etcdMonitorSystemdService":        getBase64EncodedGzippedCustomScript(etcdMonitorSystemdService, cs),
 		"healthMonitorScript":              getBase64EncodedGzippedCustomScript(kubernetesHealthMonitorScript, cs),
 		"kubeletMonitorSystemdService":     getBase64EncodedGzippedCustomScript(kubernetesKubeletMonitorSystemdService, cs),
+		"apiServerAdmissionConfiguration":  getBase64EncodedGzippedCustomScript(apiServerAdmissionConfiguration, cs),
 		"apiserverMonitorSystemdService":   getBase64EncodedGzippedCustomScript(apiserverMonitorSystemdService, cs),
 		"dockerMonitorSystemdService":      getBase64EncodedGzippedCustomScript(kubernetesDockerMonitorSystemdService, cs),
 		"provisionCIS":                     getBase64EncodedGzippedCustomScript(kubernetesCISScript, cs),
@@ -425,6 +428,7 @@ func TestK8sVars(t *testing.T) {
 		"etcdMonitorSystemdService":        getBase64EncodedGzippedCustomScript(etcdMonitorSystemdService, cs),
 		"healthMonitorScript":              getBase64EncodedGzippedCustomScript(kubernetesHealthMonitorScript, cs),
 		"kubeletMonitorSystemdService":     getBase64EncodedGzippedCustomScript(kubernetesKubeletMonitorSystemdService, cs),
+		"apiServerAdmissionConfiguration":  getBase64EncodedGzippedCustomScript(apiServerAdmissionConfiguration, cs),
 		"apiserverMonitorSystemdService":   getBase64EncodedGzippedCustomScript(apiserverMonitorSystemdService, cs),
 		"dockerMonitorSystemdService":      getBase64EncodedGzippedCustomScript(kubernetesDockerMonitorSystemdService, cs),
 		"provisionCIS":                     getBase64EncodedGzippedCustomScript(kubernetesCISScript, cs),
@@ -608,20 +612,21 @@ func TestK8sVars(t *testing.T) {
 		"primaryAvailabilitySetName":         "",
 		"primaryScaleSetName":                cs.Properties.GetPrimaryScaleSetName(),
 		"cloudInitFiles": map[string]interface{}{
-			"provisionScript":                getBase64EncodedGzippedCustomScript(kubernetesCSEMainScript, cs),
-			"provisionSource":                getBase64EncodedGzippedCustomScript(kubernetesCSEHelpersScript, cs),
-			"provisionInstalls":              getBase64EncodedGzippedCustomScript(kubernetesCSEInstall, cs),
-			"provisionConfigs":               getBase64EncodedGzippedCustomScript(kubernetesCSEConfig, cs),
-			"customSearchDomainsScript":      getBase64EncodedGzippedCustomScript(kubernetesCustomSearchDomainsScript, cs),
-			"etcdSystemdService":             getBase64EncodedGzippedCustomScript(etcdSystemdService, cs),
-			"dhcpv6ConfigurationScript":      getBase64EncodedGzippedCustomScript(dhcpv6ConfigurationScript, cs),
-			"dhcpv6SystemdService":           getBase64EncodedGzippedCustomScript(dhcpv6SystemdService, cs),
-			"kubeletSystemdService":          getBase64EncodedGzippedCustomScript(kubeletSystemdService, cs),
-			"etcdMonitorSystemdService":      getBase64EncodedGzippedCustomScript(etcdMonitorSystemdService, cs),
-			"healthMonitorScript":            getBase64EncodedGzippedCustomScript(kubernetesHealthMonitorScript, cs),
-			"kubeletMonitorSystemdService":   getBase64EncodedGzippedCustomScript(kubernetesKubeletMonitorSystemdService, cs),
-			"apiserverMonitorSystemdService": getBase64EncodedGzippedCustomScript(apiserverMonitorSystemdService, cs),
-			"dockerMonitorSystemdService":    getBase64EncodedGzippedCustomScript(kubernetesDockerMonitorSystemdService, cs),
+			"provisionScript":                 getBase64EncodedGzippedCustomScript(kubernetesCSEMainScript, cs),
+			"provisionSource":                 getBase64EncodedGzippedCustomScript(kubernetesCSEHelpersScript, cs),
+			"provisionInstalls":               getBase64EncodedGzippedCustomScript(kubernetesCSEInstall, cs),
+			"provisionConfigs":                getBase64EncodedGzippedCustomScript(kubernetesCSEConfig, cs),
+			"customSearchDomainsScript":       getBase64EncodedGzippedCustomScript(kubernetesCustomSearchDomainsScript, cs),
+			"etcdSystemdService":              getBase64EncodedGzippedCustomScript(etcdSystemdService, cs),
+			"dhcpv6ConfigurationScript":       getBase64EncodedGzippedCustomScript(dhcpv6ConfigurationScript, cs),
+			"dhcpv6SystemdService":            getBase64EncodedGzippedCustomScript(dhcpv6SystemdService, cs),
+			"kubeletSystemdService":           getBase64EncodedGzippedCustomScript(kubeletSystemdService, cs),
+			"etcdMonitorSystemdService":       getBase64EncodedGzippedCustomScript(etcdMonitorSystemdService, cs),
+			"healthMonitorScript":             getBase64EncodedGzippedCustomScript(kubernetesHealthMonitorScript, cs),
+			"kubeletMonitorSystemdService":    getBase64EncodedGzippedCustomScript(kubernetesKubeletMonitorSystemdService, cs),
+			"apiServerAdmissionConfiguration": getBase64EncodedGzippedCustomScript(apiServerAdmissionConfiguration, cs),
+			"apiserverMonitorSystemdService":  getBase64EncodedGzippedCustomScript(apiserverMonitorSystemdService, cs),
+			"dockerMonitorSystemdService":     getBase64EncodedGzippedCustomScript(kubernetesDockerMonitorSystemdService, cs),
 		},
 		"provisionConfigsCustomCloud":               getBase64EncodedGzippedCustomScript(kubernetesCSECustomCloud, cs),
 		"provisionScriptParametersCommon":           "[concat('" + cs.GetProvisionScriptParametersCommon(api.ProvisionScriptParametersInput{Location: common.WrapAsARMVariable("location"), ResourceGroup: common.WrapAsARMVariable("resourceGroup"), TenantID: common.WrapAsARMVariable("tenantID"), SubscriptionID: common.WrapAsARMVariable("subscriptionId"), ClientID: common.WrapAsARMVariable("servicePrincipalClientId"), ClientSecret: common.WrapAsARMVariable("singleQuote") + common.WrapAsARMVariable("servicePrincipalClientSecret") + common.WrapAsARMVariable("singleQuote"), APIServerCertificate: common.WrapAsParameter("apiServerCertificate"), KubeletPrivateKey: common.WrapAsParameter("clientPrivateKey"), ClusterKeyVaultName: common.WrapAsARMVariable("clusterKeyVaultName")}) + "')]",
@@ -729,21 +734,22 @@ func TestK8sVars(t *testing.T) {
 	}
 
 	expectedMap["cloudInitFiles"] = map[string]interface{}{
-		"provisionScript":                getBase64EncodedGzippedCustomScript(kubernetesCSEMainScript, cs),
-		"provisionSource":                getBase64EncodedGzippedCustomScript(kubernetesCSEHelpersScript, cs),
-		"provisionInstalls":              getBase64EncodedGzippedCustomScript(kubernetesCSEInstall, cs),
-		"provisionConfigs":               getBase64EncodedGzippedCustomScript(kubernetesCSEConfig, cs),
-		"customSearchDomainsScript":      getBase64EncodedGzippedCustomScript(kubernetesCustomSearchDomainsScript, cs),
-		"etcdSystemdService":             getBase64EncodedGzippedCustomScript(etcdSystemdService, cs),
-		"dhcpv6ConfigurationScript":      getBase64EncodedGzippedCustomScript(dhcpv6ConfigurationScript, cs),
-		"dhcpv6SystemdService":           getBase64EncodedGzippedCustomScript(dhcpv6SystemdService, cs),
-		"kubeletSystemdService":          getBase64EncodedGzippedCustomScript(kubeletSystemdService, cs),
-		"etcdMonitorSystemdService":      getBase64EncodedGzippedCustomScript(etcdMonitorSystemdService, cs),
-		"healthMonitorScript":            getBase64EncodedGzippedCustomScript(kubernetesHealthMonitorScript, cs),
-		"kubeletMonitorSystemdService":   getBase64EncodedGzippedCustomScript(kubernetesKubeletMonitorSystemdService, cs),
-		"apiserverMonitorSystemdService": getBase64EncodedGzippedCustomScript(apiserverMonitorSystemdService, cs),
-		"dockerMonitorSystemdService":    getBase64EncodedGzippedCustomScript(kubernetesDockerMonitorSystemdService, cs),
-		"systemdBPFMount":                getBase64EncodedGzippedCustomScript(systemdBPFMount, cs),
+		"provisionScript":                 getBase64EncodedGzippedCustomScript(kubernetesCSEMainScript, cs),
+		"provisionSource":                 getBase64EncodedGzippedCustomScript(kubernetesCSEHelpersScript, cs),
+		"provisionInstalls":               getBase64EncodedGzippedCustomScript(kubernetesCSEInstall, cs),
+		"provisionConfigs":                getBase64EncodedGzippedCustomScript(kubernetesCSEConfig, cs),
+		"customSearchDomainsScript":       getBase64EncodedGzippedCustomScript(kubernetesCustomSearchDomainsScript, cs),
+		"etcdSystemdService":              getBase64EncodedGzippedCustomScript(etcdSystemdService, cs),
+		"dhcpv6ConfigurationScript":       getBase64EncodedGzippedCustomScript(dhcpv6ConfigurationScript, cs),
+		"dhcpv6SystemdService":            getBase64EncodedGzippedCustomScript(dhcpv6SystemdService, cs),
+		"kubeletSystemdService":           getBase64EncodedGzippedCustomScript(kubeletSystemdService, cs),
+		"etcdMonitorSystemdService":       getBase64EncodedGzippedCustomScript(etcdMonitorSystemdService, cs),
+		"healthMonitorScript":             getBase64EncodedGzippedCustomScript(kubernetesHealthMonitorScript, cs),
+		"kubeletMonitorSystemdService":    getBase64EncodedGzippedCustomScript(kubernetesKubeletMonitorSystemdService, cs),
+		"apiServerAdmissionConfiguration": getBase64EncodedGzippedCustomScript(apiServerAdmissionConfiguration, cs),
+		"apiserverMonitorSystemdService":  getBase64EncodedGzippedCustomScript(apiserverMonitorSystemdService, cs),
+		"dockerMonitorSystemdService":     getBase64EncodedGzippedCustomScript(kubernetesDockerMonitorSystemdService, cs),
+		"systemdBPFMount":                 getBase64EncodedGzippedCustomScript(systemdBPFMount, cs),
 	}
 	expectedMap["provisionScriptParametersCommon"] = "[concat('" + cs.GetProvisionScriptParametersCommon(api.ProvisionScriptParametersInput{Location: common.WrapAsARMVariable("location"), ResourceGroup: common.WrapAsARMVariable("resourceGroup"), TenantID: common.WrapAsARMVariable("tenantID"), SubscriptionID: common.WrapAsARMVariable("subscriptionId"), ClientID: common.WrapAsARMVariable("servicePrincipalClientId"), ClientSecret: common.WrapAsARMVariable("singleQuote") + common.WrapAsARMVariable("servicePrincipalClientSecret") + common.WrapAsARMVariable("singleQuote"), APIServerCertificate: common.WrapAsParameter("apiServerCertificate"), KubeletPrivateKey: common.WrapAsParameter("clientPrivateKey"), ClusterKeyVaultName: common.WrapAsARMVariable("clusterKeyVaultName")}) + "')]"
 	diff = cmp.Diff(varMap, expectedMap)
@@ -879,20 +885,21 @@ func TestK8sVarsMastersOnly(t *testing.T) {
 		"primaryAvailabilitySetName":         "",
 		"primaryScaleSetName":                "",
 		"cloudInitFiles": map[string]interface{}{
-			"provisionScript":                getBase64EncodedGzippedCustomScript(kubernetesCSEMainScript, cs),
-			"provisionSource":                getBase64EncodedGzippedCustomScript(kubernetesCSEHelpersScript, cs),
-			"provisionInstalls":              getBase64EncodedGzippedCustomScript(kubernetesCSEInstall, cs),
-			"provisionConfigs":               getBase64EncodedGzippedCustomScript(kubernetesCSEConfig, cs),
-			"customSearchDomainsScript":      getBase64EncodedGzippedCustomScript(kubernetesCustomSearchDomainsScript, cs),
-			"etcdSystemdService":             getBase64EncodedGzippedCustomScript(etcdSystemdService, cs),
-			"dhcpv6ConfigurationScript":      getBase64EncodedGzippedCustomScript(dhcpv6ConfigurationScript, cs),
-			"dhcpv6SystemdService":           getBase64EncodedGzippedCustomScript(dhcpv6SystemdService, cs),
-			"kubeletSystemdService":          getBase64EncodedGzippedCustomScript(kubeletSystemdService, cs),
-			"etcdMonitorSystemdService":      getBase64EncodedGzippedCustomScript(etcdMonitorSystemdService, cs),
-			"healthMonitorScript":            getBase64EncodedGzippedCustomScript(kubernetesHealthMonitorScript, cs),
-			"kubeletMonitorSystemdService":   getBase64EncodedGzippedCustomScript(kubernetesKubeletMonitorSystemdService, cs),
-			"apiserverMonitorSystemdService": getBase64EncodedGzippedCustomScript(apiserverMonitorSystemdService, cs),
-			"dockerMonitorSystemdService":    getBase64EncodedGzippedCustomScript(kubernetesDockerMonitorSystemdService, cs),
+			"provisionScript":                 getBase64EncodedGzippedCustomScript(kubernetesCSEMainScript, cs),
+			"provisionSource":                 getBase64EncodedGzippedCustomScript(kubernetesCSEHelpersScript, cs),
+			"provisionInstalls":               getBase64EncodedGzippedCustomScript(kubernetesCSEInstall, cs),
+			"provisionConfigs":                getBase64EncodedGzippedCustomScript(kubernetesCSEConfig, cs),
+			"customSearchDomainsScript":       getBase64EncodedGzippedCustomScript(kubernetesCustomSearchDomainsScript, cs),
+			"etcdSystemdService":              getBase64EncodedGzippedCustomScript(etcdSystemdService, cs),
+			"dhcpv6ConfigurationScript":       getBase64EncodedGzippedCustomScript(dhcpv6ConfigurationScript, cs),
+			"dhcpv6SystemdService":            getBase64EncodedGzippedCustomScript(dhcpv6SystemdService, cs),
+			"kubeletSystemdService":           getBase64EncodedGzippedCustomScript(kubeletSystemdService, cs),
+			"etcdMonitorSystemdService":       getBase64EncodedGzippedCustomScript(etcdMonitorSystemdService, cs),
+			"healthMonitorScript":             getBase64EncodedGzippedCustomScript(kubernetesHealthMonitorScript, cs),
+			"kubeletMonitorSystemdService":    getBase64EncodedGzippedCustomScript(kubernetesKubeletMonitorSystemdService, cs),
+			"apiServerAdmissionConfiguration": getBase64EncodedGzippedCustomScript(apiServerAdmissionConfiguration, cs),
+			"apiserverMonitorSystemdService":  getBase64EncodedGzippedCustomScript(apiserverMonitorSystemdService, cs),
+			"dockerMonitorSystemdService":     getBase64EncodedGzippedCustomScript(kubernetesDockerMonitorSystemdService, cs),
 		},
 		"provisionScriptParametersCommon":           "[concat('" + cs.GetProvisionScriptParametersCommon(api.ProvisionScriptParametersInput{Location: common.WrapAsARMVariable("location"), ResourceGroup: common.WrapAsARMVariable("resourceGroup"), TenantID: common.WrapAsARMVariable("tenantID"), SubscriptionID: common.WrapAsARMVariable("subscriptionId"), ClientID: common.WrapAsARMVariable("servicePrincipalClientId"), ClientSecret: common.WrapAsARMVariable("singleQuote") + common.WrapAsARMVariable("servicePrincipalClientSecret") + common.WrapAsARMVariable("singleQuote"), APIServerCertificate: common.WrapAsParameter("apiServerCertificate"), KubeletPrivateKey: common.WrapAsParameter("clientPrivateKey"), ClusterKeyVaultName: common.WrapAsARMVariable("clusterKeyVaultName")}) + "')]",
 		"provisionScriptParametersMaster":           "[concat('COSMOS_URI= MASTER_VM_NAME=',variables('masterVMNames')[variables('masterOffset')],' ETCD_PEER_URL=',variables('masterEtcdPeerURLs')[variables('masterOffset')],' ETCD_CLIENT_URL=',variables('masterEtcdClientURLs')[variables('masterOffset')],' MASTER_NODE=true NO_OUTBOUND=false AUDITD_ENABLED=false CLUSTER_AUTOSCALER_ADDON=false APISERVER_PRIVATE_KEY=',parameters('apiServerPrivateKey'),' CA_CERTIFICATE=',parameters('caCertificate'),' CA_PRIVATE_KEY=',parameters('caPrivateKey'),' MASTER_FQDN=',variables('masterFqdnPrefix'),' KUBECONFIG_CERTIFICATE=',parameters('kubeConfigCertificate'),' KUBECONFIG_KEY=',parameters('kubeConfigPrivateKey'),' ETCD_SERVER_CERTIFICATE=',parameters('etcdServerCertificate'),' ETCD_CLIENT_CERTIFICATE=',parameters('etcdClientCertificate'),' ETCD_SERVER_PRIVATE_KEY=',parameters('etcdServerPrivateKey'),' ETCD_CLIENT_PRIVATE_KEY=',parameters('etcdClientPrivateKey'),' ETCD_PEER_CERTIFICATES=',string(variables('etcdPeerCertificates')),' ETCD_PEER_PRIVATE_KEYS=',string(variables('etcdPeerPrivateKeys')),' ENABLE_AGGREGATED_APIS=',string(parameters('enableAggregatedAPIs')),' KUBECONFIG_SERVER=',variables('kubeconfigServer'))]",
@@ -941,22 +948,23 @@ func TestK8sVarsMastersOnly(t *testing.T) {
 	cs.Properties.OrchestratorProfile.KubernetesConfig.EnableEncryptionWithExternalKms = to.BoolPtr(true)
 	expectedMap["clusterKeyVaultName"] = string("[take(concat('kv', tolower(uniqueString(concat(variables('masterFqdnPrefix'),variables('location'),parameters('nameSuffix'))))), 22)]")
 	expectedMap["cloudInitFiles"] = map[string]interface{}{
-		"provisionScript":                getBase64EncodedGzippedCustomScript(kubernetesCSEMainScript, cs),
-		"provisionSource":                getBase64EncodedGzippedCustomScript(kubernetesCSEHelpersScript, cs),
-		"provisionInstalls":              getBase64EncodedGzippedCustomScript(kubernetesCSEInstall, cs),
-		"provisionConfigs":               getBase64EncodedGzippedCustomScript(kubernetesCSEConfig, cs),
-		"customSearchDomainsScript":      getBase64EncodedGzippedCustomScript(kubernetesCustomSearchDomainsScript, cs),
-		"etcdSystemdService":             getBase64EncodedGzippedCustomScript(etcdSystemdService, cs),
-		"dhcpv6ConfigurationScript":      getBase64EncodedGzippedCustomScript(dhcpv6ConfigurationScript, cs),
-		"dhcpv6SystemdService":           getBase64EncodedGzippedCustomScript(dhcpv6SystemdService, cs),
-		"kubeletSystemdService":          getBase64EncodedGzippedCustomScript(kubeletSystemdService, cs),
-		"etcdMonitorSystemdService":      getBase64EncodedGzippedCustomScript(etcdMonitorSystemdService, cs),
-		"healthMonitorScript":            getBase64EncodedGzippedCustomScript(kubernetesHealthMonitorScript, cs),
-		"kubeletMonitorSystemdService":   getBase64EncodedGzippedCustomScript(kubernetesKubeletMonitorSystemdService, cs),
-		"apiserverMonitorSystemdService": getBase64EncodedGzippedCustomScript(apiserverMonitorSystemdService, cs),
-		"dockerMonitorSystemdService":    getBase64EncodedGzippedCustomScript(kubernetesDockerMonitorSystemdService, cs),
-		"kmsKeyvaultKeySystemdService":   getBase64EncodedGzippedCustomScript(kmsKeyvaultKeySystemdService, cs),
-		"kmsKeyvaultKeyScript":           getBase64EncodedGzippedCustomScript(kmsKeyvaultKeyScript, cs),
+		"provisionScript":                 getBase64EncodedGzippedCustomScript(kubernetesCSEMainScript, cs),
+		"provisionSource":                 getBase64EncodedGzippedCustomScript(kubernetesCSEHelpersScript, cs),
+		"provisionInstalls":               getBase64EncodedGzippedCustomScript(kubernetesCSEInstall, cs),
+		"provisionConfigs":                getBase64EncodedGzippedCustomScript(kubernetesCSEConfig, cs),
+		"customSearchDomainsScript":       getBase64EncodedGzippedCustomScript(kubernetesCustomSearchDomainsScript, cs),
+		"etcdSystemdService":              getBase64EncodedGzippedCustomScript(etcdSystemdService, cs),
+		"dhcpv6ConfigurationScript":       getBase64EncodedGzippedCustomScript(dhcpv6ConfigurationScript, cs),
+		"dhcpv6SystemdService":            getBase64EncodedGzippedCustomScript(dhcpv6SystemdService, cs),
+		"kubeletSystemdService":           getBase64EncodedGzippedCustomScript(kubeletSystemdService, cs),
+		"etcdMonitorSystemdService":       getBase64EncodedGzippedCustomScript(etcdMonitorSystemdService, cs),
+		"healthMonitorScript":             getBase64EncodedGzippedCustomScript(kubernetesHealthMonitorScript, cs),
+		"kubeletMonitorSystemdService":    getBase64EncodedGzippedCustomScript(kubernetesKubeletMonitorSystemdService, cs),
+		"apiserverMonitorSystemdService":  getBase64EncodedGzippedCustomScript(apiserverMonitorSystemdService, cs),
+		"dockerMonitorSystemdService":     getBase64EncodedGzippedCustomScript(kubernetesDockerMonitorSystemdService, cs),
+		"kmsKeyvaultKeySystemdService":    getBase64EncodedGzippedCustomScript(kmsKeyvaultKeySystemdService, cs),
+		"kmsKeyvaultKeyScript":            getBase64EncodedGzippedCustomScript(kmsKeyvaultKeyScript, cs),
+		"apiServerAdmissionConfiguration": getBase64EncodedGzippedCustomScript(apiServerAdmissionConfiguration, cs),
 	}
 
 	varMap, err = GetKubernetesVariables(cs)
