@@ -1756,7 +1756,7 @@ func translationsIt_itLc_messagesAcsenginePo() (*asset, error) {
 	return a, nil
 }
 
-var _translationsJa_jpLc_messagesAcsengineMo = []byte("\xde\x12\x04\x95\x00\x00\x00\x00\x15\x00\x00\x00\x1c\x00\x00\x00\xc4\x00\x00\x00\x1d\x00\x00\x00l\x01\x00\x00\x00\x00\x00\x00\xe0\x01\x00\x00 \x00\x00\x00\xe1\x01\x00\x00+\x00\x00\x00\x02\x02\x00\x00F\x00\x00\x00.\x02\x00\x00I\x00\x00\x00u\x02\x00\x009\x00\x00\x00\xbf\x02\x00\x00=\x00\x00\x00\xf9\x02\x00\x00C\x00\x00\x007\x03\x00\x00!\x00\x00\x00{\x03\x00\x00(\x00\x00\x00\x9d\x03\x00\x00%\x00\x00\x00\xc6\x03\x00\x00\x19\x00\x00\x00\xec\x03\x00\x00\x19\x00\x00\x00\x06\x04\x00\x00.\x00\x00\x00 \x04\x00\x00*\x00\x00\x00O\x04\x00\x00+\x00\x00\x00z\x04\x00\x00>\x00\x00\x00\xa6\x04\x00\x00 \x00\x00\x00\xe5\x04\x00\x00\x1f\x00\x00\x00\x06\x05\x00\x00\x1c\x00\x00\x00&\x05\x00\x00\x1b\x00\x00\x00C\x05\x00\x00\x84\x01\x00\x00_\x05\x00\x00S\x00\x00\x00\xe4\x06\x00\x00E\x00\x00\x008\a\x00\x00{\x00\x00\x00~\a\x00\x00x\x00\x00\x00\xfa\a\x00\x00Q\x00\x00\x00s\b\x00\x00r\x00\x00\x00\xc5\b\x00\x00\u007f\x00\x00\x008\t\x00\x00I\x00\x00\x00\xb8\t\x00\x00V\x00\x00\x00\x02\n\x00\x00Y\x00\x00\x00Y\n\x00\x00A\x00\x00\x00\xb3\n\x00\x00G\x00\x00\x00\xf5\n\x00\x00E\x00\x00\x00=\v\x00\x004\x00\x00\x00\x83\v\x00\x008\x00\x00\x00\xb8\v\x00\x00m\x00\x00\x00\xf1\v\x00\x00E\x00\x00\x00_\f\x00\x008\x00\x00\x00\xa5\f\x00\x00)\x00\x00\x00\xde\f\x00\x00*\x00\x00\x00\b\r\x00\x00\x01\x00\x00\x00\f\x00\x00\x00\x02\x00\x00\x00\x04\x00\x00\x00\x06\x00\x00\x00\a\x00\x00\x00\t\x00\x00\x00\x0f\x00\x00\x00\x12\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\n\x00\x00\x00\x00\x00\x00\x00\x14\x00\x00\x00\x11\x00\x00\x00\b\x00\x00\x00\x13\x00\x00\x00\x0e\x00\x00\x00\x05\x00\x00\x00\x10\x00\x00\x00\x03\x00\x00\x00\x15\x00\x00\x00\x00\x00\x00\x00\v\x00\x00\x00\r\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00Error reading file %s, Error: %s\x00Error while querying ARM for resources: %+v\x00Found 2 resources with type %s in the template. There should only be 1\x00Found no resources with type %s in the template. There should have been 1\x00Total count of master VMs: %d exceeded expected count: %d\x00Upgrade to Kubernetes 1.6.2 is not supported from version: %s\x00Upgrade to Kubernetes version: %s is not supported from version: %s\x00error creating directory '%s': %s\x00error executing template for file %s: %v\x00error generating upgrade template: %s\x00error parsing file %s: %v\x00error reading file %s: %s\x00failed to create openssh public key string: %q\x00failed to generate private key for ssh: %q\x00failed to initialize template generator: %s\x00invalid version %s for conversion back from unversioned object\x00orchestrator '%s' is unsupported\x00template file %s does not exist\x00unrecognized APIVersion '%s'\x00yaml file %s does not exist\x00Project-Id-Version: acsengine\nReport-Msgid-Bugs-To: \nPOT-Creation-Date: 2017-07-25 00:04+0000\nPO-Revision-Date: 2017-07-24 17:23-0700\nLast-Translator: Jiangtian Li <jiangtianli@hotmail.com>\nLanguage-Team: English\nLanguage: en_US\nMIME-Version: 1.0\nContent-Type: text/plain; charset=UTF-8\nContent-Transfer-Encoding: 8bit\nPlural-Forms: nplurals=2; plural=(n != 1);\nX-Generator: Poedit 2.0.3\n\x00\u30d5\u30a1\u30a4\u30eb %s \u306e\u8aad\u307f\u53d6\u308a\u4e2d\u306b\u30a8\u30e9\u30fc\u304c\u767a\u751f\u3057\u307e\u3057\u305f\u3002\u30a8\u30e9\u30fc: %s\x00ARM \u306b\u30ea\u30bd\u30fc\u30b9\u3092\u7167\u4f1a\u4e2d\u306b\u30a8\u30e9\u30fc\u304c\u767a\u751f\u3057\u307e\u3057\u305f: %+v\x00\u30c6\u30f3\u30d7\u30ec\u30fc\u30c8\u306b\u7a2e\u985e %s \u306e\u30ea\u30bd\u30fc\u30b9\u304c 2 \u3064\u898b\u3064\u304b\u308a\u307e\u3057\u305f\u30021 \u3064\u3060\u3051\u898b\u3064\u304b\u308b\u5fc5\u8981\u304c\u3042\u308a\u307e\u3059\x00\u30c6\u30f3\u30d7\u30ec\u30fc\u30c8\u306b\u7a2e\u985e %s \u306e\u30ea\u30bd\u30fc\u30b9\u304c\u898b\u3064\u304b\u308a\u307e\u305b\u3093\u3067\u3057\u305f\u30021 \u3064\u898b\u3064\u304b\u308b\u5fc5\u8981\u304c\u3042\u308a\u307e\u3059\x00\u30de\u30b9\u30bf\u30fc VM \u306e\u5408\u8a08\u6570 %d \u304c\u3001\u60f3\u5b9a\u3055\u308c\u308b\u6570 %d \u3092\u8d85\u3048\u3066\u3044\u307e\u3059\x00Kubernetes 1.6.2 \u306b\u30d0\u30fc\u30b8\u30e7\u30f3 %s \u304b\u3089\u30a2\u30c3\u30d7\u30b0\u30ec\u30fc\u30c9\u3059\u308b\u3053\u3068\u306f\u30b5\u30dd\u30fc\u30c8\u3055\u308c\u3066\u3044\u307e\u305b\u3093\x00Kubernetes \u30d0\u30fc\u30b8\u30e7\u30f3 %s \u306b\u30d0\u30fc\u30b8\u30e7\u30f3 %s \u304b\u3089\u30a2\u30c3\u30d7\u30b0\u30ec\u30fc\u30c9\u3059\u308b\u3053\u3068\u306f\u30b5\u30dd\u30fc\u30c8\u3055\u308c\u3066\u3044\u307e\u305b\u3093\x00\u30c7\u30a3\u30ec\u30af\u30c8\u30ea '%s' \u306e\u4f5c\u6210\u4e2d\u306b\u30a8\u30e9\u30fc\u304c\u767a\u751f\u3057\u307e\u3057\u305f: %s\x00\u30d5\u30a1\u30a4\u30eb %s \u306b\u30c6\u30f3\u30d7\u30ec\u30fc\u30c8\u3092\u5b9f\u884c\u4e2d\u306b\u30a8\u30e9\u30fc\u304c\u767a\u751f\u3057\u307e\u3057\u305f: %v\x00\u30a2\u30c3\u30d7\u30b0\u30ec\u30fc\u30c9 \u30c6\u30f3\u30d7\u30ec\u30fc\u30c8\u306e\u751f\u6210\u4e2d\u306b\u30a8\u30e9\u30fc\u304c\u767a\u751f\u3057\u307e\u3057\u305f: %s\x00\u30d5\u30a1\u30a4\u30eb %s \u306e\u89e3\u6790\u4e2d\u306b\u30a8\u30e9\u30fc\u304c\u767a\u751f\u3057\u307e\u3057\u305f: %v\x00\u30d5\u30a1\u30a4\u30eb %s \u306e\u8aad\u307f\u53d6\u308a\u4e2d\u306b\u30a8\u30e9\u30fc\u304c\u767a\u751f\u3057\u307e\u3057\u305f: %s\x00OpenSSH \u516c\u958b\u30ad\u30fc\u306e\u6587\u5b57\u5217\u3092\u4f5c\u6210\u3067\u304d\u307e\u305b\u3093\u3067\u3057\u305f: %q\x00SSH \u306e\u79d8\u5bc6\u30ad\u30fc\u3092\u751f\u6210\u3067\u304d\u307e\u305b\u3093\u3067\u3057\u305f\x00Template Generator \u306e\u521d\u671f\u5316\u306b\u5931\u6557\u3057\u307e\u3057\u305f: %s\x00\u30d0\u30fc\u30b8\u30e7\u30f3\u7ba1\u7406\u5916\u30aa\u30d6\u30b8\u30a7\u30af\u30c8\u304b\u3089\u5909\u63db\u3057\u76f4\u3059\u3053\u3068\u306e\u3067\u304d\u306a\u3044\u30d0\u30fc\u30b8\u30e7\u30f3\u3067\u3059: %s\x00\u30aa\u30fc\u30b1\u30b9\u30c8\u30ec\u30fc\u30bf\u30fc '%s' \u306f\u30b5\u30dd\u30fc\u30c8\u3055\u308c\u3066\u3044\u307e\u305b\u3093\x00\u30c6\u30f3\u30d7\u30ec\u30fc\u30c8 \u30d5\u30a1\u30a4\u30eb %s \u304c\u5b58\u5728\u3057\u307e\u305b\u3093\x00\u8a8d\u8b58\u3055\u308c\u306a\u3044 APIVersion\u3067\u3059: '%s'\x00YAML \u30d5\u30a1\u30a4\u30eb %s \u304c\u5b58\u5728\u3057\u307e\u305b\u3093\x00")
+var _translationsJa_jpLc_messagesAcsengineMo = []byte("\xde\x12\x04\x95\x00\x00\x00\x00\x15\x00\x00\x00\x1c\x00\x00\x00\xc4\x00\x00\x00\x1d\x00\x00\x00l\x01\x00\x00\x00\x00\x00\x00\xe0\x01\x00\x00 \x00\x00\x00\xe1\x01\x00\x00+\x00\x00\x00\x02\x02\x00\x00F\x00\x00\x00.\x02\x00\x00I\x00\x00\x00u\x02\x00\x009\x00\x00\x00\xbf\x02\x00\x00=\x00\x00\x00\xf9\x02\x00\x00C\x00\x00\x007\x03\x00\x00!\x00\x00\x00{\x03\x00\x00(\x00\x00\x00\x9d\x03\x00\x00%\x00\x00\x00\xc6\x03\x00\x00\x19\x00\x00\x00\xec\x03\x00\x00\x19\x00\x00\x00\x06\x04\x00\x00.\x00\x00\x00 \x04\x00\x00*\x00\x00\x00O\x04\x00\x00+\x00\x00\x00z\x04\x00\x00>\x00\x00\x00\xa6\x04\x00\x00 \x00\x00\x00\xe5\x04\x00\x00\x1f\x00\x00\x00\x06\x05\x00\x00\x1c\x00\x00\x00&\x05\x00\x00\x1b\x00\x00\x00C\x05\x00\x00\x84\x01\x00\x00_\x05\x00\x00S\x00\x00\x00\xe4\x06\x00\x00E\x00\x00\x008\a\x00\x00{\x00\x00\x00~\a\x00\x00x\x00\x00\x00\xfa\a\x00\x00Q\x00\x00\x00s\b\x00\x00r\x00\x00\x00\xc5\b\x00\x00\x7f\x00\x00\x008\t\x00\x00I\x00\x00\x00\xb8\t\x00\x00V\x00\x00\x00\x02\n\x00\x00Y\x00\x00\x00Y\n\x00\x00A\x00\x00\x00\xb3\n\x00\x00G\x00\x00\x00\xf5\n\x00\x00E\x00\x00\x00=\v\x00\x004\x00\x00\x00\x83\v\x00\x008\x00\x00\x00\xb8\v\x00\x00m\x00\x00\x00\xf1\v\x00\x00E\x00\x00\x00_\f\x00\x008\x00\x00\x00\xa5\f\x00\x00)\x00\x00\x00\xde\f\x00\x00*\x00\x00\x00\b\r\x00\x00\x01\x00\x00\x00\f\x00\x00\x00\x02\x00\x00\x00\x04\x00\x00\x00\x06\x00\x00\x00\a\x00\x00\x00\t\x00\x00\x00\x0f\x00\x00\x00\x12\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\n\x00\x00\x00\x00\x00\x00\x00\x14\x00\x00\x00\x11\x00\x00\x00\b\x00\x00\x00\x13\x00\x00\x00\x0e\x00\x00\x00\x05\x00\x00\x00\x10\x00\x00\x00\x03\x00\x00\x00\x15\x00\x00\x00\x00\x00\x00\x00\v\x00\x00\x00\r\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00Error reading file %s, Error: %s\x00Error while querying ARM for resources: %+v\x00Found 2 resources with type %s in the template. There should only be 1\x00Found no resources with type %s in the template. There should have been 1\x00Total count of master VMs: %d exceeded expected count: %d\x00Upgrade to Kubernetes 1.6.2 is not supported from version: %s\x00Upgrade to Kubernetes version: %s is not supported from version: %s\x00error creating directory '%s': %s\x00error executing template for file %s: %v\x00error generating upgrade template: %s\x00error parsing file %s: %v\x00error reading file %s: %s\x00failed to create openssh public key string: %q\x00failed to generate private key for ssh: %q\x00failed to initialize template generator: %s\x00invalid version %s for conversion back from unversioned object\x00orchestrator '%s' is unsupported\x00template file %s does not exist\x00unrecognized APIVersion '%s'\x00yaml file %s does not exist\x00Project-Id-Version: acsengine\nReport-Msgid-Bugs-To: \nPOT-Creation-Date: 2017-07-25 00:04+0000\nPO-Revision-Date: 2017-07-24 17:23-0700\nLast-Translator: Jiangtian Li <jiangtianli@hotmail.com>\nLanguage-Team: English\nLanguage: en_US\nMIME-Version: 1.0\nContent-Type: text/plain; charset=UTF-8\nContent-Transfer-Encoding: 8bit\nPlural-Forms: nplurals=2; plural=(n != 1);\nX-Generator: Poedit 2.0.3\n\x00\u30d5\u30a1\u30a4\u30eb %s \u306e\u8aad\u307f\u53d6\u308a\u4e2d\u306b\u30a8\u30e9\u30fc\u304c\u767a\u751f\u3057\u307e\u3057\u305f\u3002\u30a8\u30e9\u30fc: %s\x00ARM \u306b\u30ea\u30bd\u30fc\u30b9\u3092\u7167\u4f1a\u4e2d\u306b\u30a8\u30e9\u30fc\u304c\u767a\u751f\u3057\u307e\u3057\u305f: %+v\x00\u30c6\u30f3\u30d7\u30ec\u30fc\u30c8\u306b\u7a2e\u985e %s \u306e\u30ea\u30bd\u30fc\u30b9\u304c 2 \u3064\u898b\u3064\u304b\u308a\u307e\u3057\u305f\u30021 \u3064\u3060\u3051\u898b\u3064\u304b\u308b\u5fc5\u8981\u304c\u3042\u308a\u307e\u3059\x00\u30c6\u30f3\u30d7\u30ec\u30fc\u30c8\u306b\u7a2e\u985e %s \u306e\u30ea\u30bd\u30fc\u30b9\u304c\u898b\u3064\u304b\u308a\u307e\u305b\u3093\u3067\u3057\u305f\u30021 \u3064\u898b\u3064\u304b\u308b\u5fc5\u8981\u304c\u3042\u308a\u307e\u3059\x00\u30de\u30b9\u30bf\u30fc VM \u306e\u5408\u8a08\u6570 %d \u304c\u3001\u60f3\u5b9a\u3055\u308c\u308b\u6570 %d \u3092\u8d85\u3048\u3066\u3044\u307e\u3059\x00Kubernetes 1.6.2 \u306b\u30d0\u30fc\u30b8\u30e7\u30f3 %s \u304b\u3089\u30a2\u30c3\u30d7\u30b0\u30ec\u30fc\u30c9\u3059\u308b\u3053\u3068\u306f\u30b5\u30dd\u30fc\u30c8\u3055\u308c\u3066\u3044\u307e\u305b\u3093\x00Kubernetes \u30d0\u30fc\u30b8\u30e7\u30f3 %s \u306b\u30d0\u30fc\u30b8\u30e7\u30f3 %s \u304b\u3089\u30a2\u30c3\u30d7\u30b0\u30ec\u30fc\u30c9\u3059\u308b\u3053\u3068\u306f\u30b5\u30dd\u30fc\u30c8\u3055\u308c\u3066\u3044\u307e\u305b\u3093\x00\u30c7\u30a3\u30ec\u30af\u30c8\u30ea '%s' \u306e\u4f5c\u6210\u4e2d\u306b\u30a8\u30e9\u30fc\u304c\u767a\u751f\u3057\u307e\u3057\u305f: %s\x00\u30d5\u30a1\u30a4\u30eb %s \u306b\u30c6\u30f3\u30d7\u30ec\u30fc\u30c8\u3092\u5b9f\u884c\u4e2d\u306b\u30a8\u30e9\u30fc\u304c\u767a\u751f\u3057\u307e\u3057\u305f: %v\x00\u30a2\u30c3\u30d7\u30b0\u30ec\u30fc\u30c9 \u30c6\u30f3\u30d7\u30ec\u30fc\u30c8\u306e\u751f\u6210\u4e2d\u306b\u30a8\u30e9\u30fc\u304c\u767a\u751f\u3057\u307e\u3057\u305f: %s\x00\u30d5\u30a1\u30a4\u30eb %s \u306e\u89e3\u6790\u4e2d\u306b\u30a8\u30e9\u30fc\u304c\u767a\u751f\u3057\u307e\u3057\u305f: %v\x00\u30d5\u30a1\u30a4\u30eb %s \u306e\u8aad\u307f\u53d6\u308a\u4e2d\u306b\u30a8\u30e9\u30fc\u304c\u767a\u751f\u3057\u307e\u3057\u305f: %s\x00OpenSSH \u516c\u958b\u30ad\u30fc\u306e\u6587\u5b57\u5217\u3092\u4f5c\u6210\u3067\u304d\u307e\u305b\u3093\u3067\u3057\u305f: %q\x00SSH \u306e\u79d8\u5bc6\u30ad\u30fc\u3092\u751f\u6210\u3067\u304d\u307e\u305b\u3093\u3067\u3057\u305f\x00Template Generator \u306e\u521d\u671f\u5316\u306b\u5931\u6557\u3057\u307e\u3057\u305f: %s\x00\u30d0\u30fc\u30b8\u30e7\u30f3\u7ba1\u7406\u5916\u30aa\u30d6\u30b8\u30a7\u30af\u30c8\u304b\u3089\u5909\u63db\u3057\u76f4\u3059\u3053\u3068\u306e\u3067\u304d\u306a\u3044\u30d0\u30fc\u30b8\u30e7\u30f3\u3067\u3059: %s\x00\u30aa\u30fc\u30b1\u30b9\u30c8\u30ec\u30fc\u30bf\u30fc '%s' \u306f\u30b5\u30dd\u30fc\u30c8\u3055\u308c\u3066\u3044\u307e\u305b\u3093\x00\u30c6\u30f3\u30d7\u30ec\u30fc\u30c8 \u30d5\u30a1\u30a4\u30eb %s \u304c\u5b58\u5728\u3057\u307e\u305b\u3093\x00\u8a8d\u8b58\u3055\u308c\u306a\u3044 APIVersion\u3067\u3059: '%s'\x00YAML \u30d5\u30a1\u30a4\u30eb %s \u304c\u5b58\u5728\u3057\u307e\u305b\u3093\x00")
 
 func translationsJa_jpLc_messagesAcsengineMoBytes() ([]byte, error) {
 	return _translationsJa_jpLc_messagesAcsengineMo, nil
@@ -2712,7 +2712,7 @@ func translationsPt_ptLc_messagesAcsenginePo() (*asset, error) {
 	return a, nil
 }
 
-var _translationsRu_ruLc_messagesAcsengineMo = []byte("\xde\x12\x04\x95\x00\x00\x00\x00\x15\x00\x00\x00\x1c\x00\x00\x00\xc4\x00\x00\x00\x1d\x00\x00\x00l\x01\x00\x00\x00\x00\x00\x00\xe0\x01\x00\x00 \x00\x00\x00\xe1\x01\x00\x00+\x00\x00\x00\x02\x02\x00\x00F\x00\x00\x00.\x02\x00\x00I\x00\x00\x00u\x02\x00\x009\x00\x00\x00\xbf\x02\x00\x00=\x00\x00\x00\xf9\x02\x00\x00C\x00\x00\x007\x03\x00\x00!\x00\x00\x00{\x03\x00\x00(\x00\x00\x00\x9d\x03\x00\x00%\x00\x00\x00\xc6\x03\x00\x00\x19\x00\x00\x00\xec\x03\x00\x00\x19\x00\x00\x00\x06\x04\x00\x00.\x00\x00\x00 \x04\x00\x00*\x00\x00\x00O\x04\x00\x00+\x00\x00\x00z\x04\x00\x00>\x00\x00\x00\xa6\x04\x00\x00 \x00\x00\x00\xe5\x04\x00\x00\x1f\x00\x00\x00\x06\x05\x00\x00\x1c\x00\x00\x00&\x05\x00\x00\x1b\x00\x00\x00C\x05\x00\x00\x84\x01\x00\x00_\x05\x00\x00@\x00\x00\x00\xe4\x06\x00\x00?\x00\x00\x00%\a\x00\x00\u007f\x00\x00\x00e\a\x00\x00\x8c\x00\x00\x00\xe5\a\x00\x00y\x00\x00\x00r\b\x00\x00a\x00\x00\x00\xec\b\x00\x00^\x00\x00\x00N\t\x00\x00>\x00\x00\x00\xad\t\x00\x00P\x00\x00\x00\xec\t\x00\x00L\x00\x00\x00=\n\x00\x004\x00\x00\x00\x8a\n\x00\x002\x00\x00\x00\xbf\n\x00\x00Y\x00\x00\x00\xf2\n\x00\x00K\x00\x00\x00L\v\x00\x00\\\x00\x00\x00\x98\v\x00\x00\x87\x00\x00\x00\xf5\v\x00\x00=\x00\x00\x00}\f\x00\x004\x00\x00\x00\xbb\f\x00\x00=\x00\x00\x00\xf0\f\x00\x00*\x00\x00\x00.\r\x00\x00\x01\x00\x00\x00\f\x00\x00\x00\x02\x00\x00\x00\x04\x00\x00\x00\x06\x00\x00\x00\a\x00\x00\x00\t\x00\x00\x00\x0f\x00\x00\x00\x12\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\n\x00\x00\x00\x00\x00\x00\x00\x14\x00\x00\x00\x11\x00\x00\x00\b\x00\x00\x00\x13\x00\x00\x00\x0e\x00\x00\x00\x05\x00\x00\x00\x10\x00\x00\x00\x03\x00\x00\x00\x15\x00\x00\x00\x00\x00\x00\x00\v\x00\x00\x00\r\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00Error reading file %s, Error: %s\x00Error while querying ARM for resources: %+v\x00Found 2 resources with type %s in the template. There should only be 1\x00Found no resources with type %s in the template. There should have been 1\x00Total count of master VMs: %d exceeded expected count: %d\x00Upgrade to Kubernetes 1.6.2 is not supported from version: %s\x00Upgrade to Kubernetes version: %s is not supported from version: %s\x00error creating directory '%s': %s\x00error executing template for file %s: %v\x00error generating upgrade template: %s\x00error parsing file %s: %v\x00error reading file %s: %s\x00failed to create openssh public key string: %q\x00failed to generate private key for ssh: %q\x00failed to initialize template generator: %s\x00invalid version %s for conversion back from unversioned object\x00orchestrator '%s' is unsupported\x00template file %s does not exist\x00unrecognized APIVersion '%s'\x00yaml file %s does not exist\x00Project-Id-Version: acsengine\nReport-Msgid-Bugs-To: \nPOT-Creation-Date: 2017-07-25 00:04+0000\nPO-Revision-Date: 2017-07-24 17:23-0700\nLast-Translator: Jiangtian Li <jiangtianli@hotmail.com>\nLanguage-Team: English\nLanguage: en_US\nMIME-Version: 1.0\nContent-Type: text/plain; charset=UTF-8\nContent-Transfer-Encoding: 8bit\nPlural-Forms: nplurals=2; plural=(n != 1);\nX-Generator: Poedit 2.0.3\n\x00\u041e\u0448\u0438\u0431\u043a\u0430 \u043f\u0440\u0438 \u0447\u0442\u0435\u043d\u0438\u0438 \u0444\u0430\u0439\u043b\u0430 %s, \u043e\u0448\u0438\u0431\u043a\u0430: %s\x00\u041e\u0448\u0438\u0431\u043a\u0430 \u043f\u0440\u0438 \u0437\u0430\u043f\u0440\u043e\u0441\u0435 \u0440\u0435\u0441\u0443\u0440\u0441\u043e\u0432 \u0432 ARM: %+v\x00\u0412 \u0448\u0430\u0431\u043b\u043e\u043d\u0435 \u043d\u0430\u0439\u0434\u0435\u043d\u043e 2\u00a0\u0440\u0435\u0441\u0443\u0440\u0441\u0430 \u0441 \u0442\u0438\u043f\u043e\u043c %s. \u0422\u0430\u043a\u043e\u0439 \u0440\u0435\u0441\u0443\u0440\u0441 \u0434\u043e\u043b\u0436\u0435\u043d \u0431\u044b\u0442\u044c \u0432\u0441\u0435\u0433\u043e 1\x00\u0412 \u0448\u0430\u0431\u043b\u043e\u043d\u0435 \u043d\u0435 \u043d\u0430\u0439\u0434\u0435\u043d\u043e \u0440\u0435\u0441\u0443\u0440\u0441\u043e\u0432 \u0441 \u0442\u0438\u043f\u043e\u043c %s. \u0414\u043e\u043b\u0436\u0435\u043d \u043f\u0440\u0438\u0441\u0443\u0442\u0441\u0442\u0432\u043e\u0432\u0430\u0442\u044c 1 \u0442\u0430\u043a\u043e\u0439 \u0440\u0435\u0441\u0443\u0440\u0441\x00\u0427\u0438\u0441\u043b\u043e \u0433\u043b\u0430\u0432\u043d\u044b\u0445 \u0432\u0438\u0440\u0442\u0443\u0430\u043b\u044c\u043d\u044b\u0445 \u043c\u0430\u0448\u0438\u043d %d \u043f\u0440\u0435\u0432\u044b\u0441\u0438\u043b\u043e \u043e\u0436\u0438\u0434\u0430\u0435\u043c\u043e\u0435 \u0437\u043d\u0430\u0447\u0435\u043d\u0438\u0435: %d\x00\u041d\u0435 \u043f\u043e\u0434\u0434\u0435\u0440\u0436\u0438\u0432\u0430\u0435\u0442\u0441\u044f \u043e\u0431\u043d\u043e\u0432\u043b\u0435\u043d\u0438\u0435 \u0434\u043e Kubernetes\u00a01.6.2 \u0441 \u0432\u0435\u0440\u0441\u0438\u0438: %s\x00\u041d\u0435 \u043f\u043e\u0434\u0434\u0435\u0440\u0436\u0438\u0432\u0430\u0435\u0442\u0441\u044f \u043e\u0431\u043d\u043e\u0432\u043b\u0435\u043d\u0438\u0435 \u0434\u043e Kubernetes\u00a0%s \u0441 \u0432\u0435\u0440\u0441\u0438\u0438: %s\x00\u043e\u0448\u0438\u0431\u043a\u0430 \u043f\u0440\u0438 \u0441\u043e\u0437\u0434\u0430\u043d\u0438\u0438 \u043a\u0430\u0442\u0430\u043b\u043e\u0433\u0430 \"%s\": %s\x00\u043e\u0448\u0438\u0431\u043a\u0430 \u043f\u0440\u0438 \u0432\u044b\u043f\u043e\u043b\u043d\u0435\u043d\u0438\u0438 \u0448\u0430\u0431\u043b\u043e\u043d\u0430 \u0434\u043b\u044f \u0444\u0430\u0439\u043b\u0430 %s: %v\x00\u043e\u0448\u0438\u0431\u043a\u0430 \u043f\u0440\u0438 \u0441\u043e\u0437\u0434\u0430\u043d\u0438\u0438 \u0448\u0430\u0431\u043b\u043e\u043d\u0430 \u043e\u0431\u043d\u043e\u0432\u043b\u0435\u043d\u0438\u044f: %s\x00\u043e\u0448\u0438\u0431\u043a\u0430 \u043f\u0440\u0438 \u0430\u043d\u0430\u043b\u0438\u0437\u0435 \u0444\u0430\u0439\u043b\u0430 %s: %v\x00\u043e\u0448\u0438\u0431\u043a\u0430 \u043f\u0440\u0438 \u0447\u0442\u0435\u043d\u0438\u0438 \u0444\u0430\u0439\u043b\u0430 %s: %s\x00\u043d\u0435 \u0443\u0434\u0430\u043b\u043e\u0441\u044c \u0441\u043e\u0437\u0434\u0430\u0442\u044c \u0441\u0442\u0440\u043e\u043a\u0443 \u043e\u0442\u043a\u0440\u044b\u0442\u043e\u0433\u043e \u043a\u043b\u044e\u0447\u0430 OpenSSH: %q\x00\u043d\u0435 \u0443\u0434\u0430\u043b\u043e\u0441\u044c \u0441\u043e\u0437\u0434\u0430\u0442\u044c \u0437\u0430\u043a\u0440\u044b\u0442\u044b\u0439 \u043a\u043b\u044e\u0447 \u0434\u043b\u044f SSH: %q\x00\u043d\u0435 \u0443\u0434\u0430\u043b\u043e\u0441\u044c \u0438\u043d\u0438\u0446\u0438\u0430\u043b\u0438\u0437\u0438\u0440\u043e\u0432\u0430\u0442\u044c \u0433\u0435\u043d\u0435\u0440\u0430\u0442\u043e\u0440 \u0448\u0430\u0431\u043b\u043e\u043d\u043e\u0432: %s\x00\u043d\u0435\u0434\u043e\u043f\u0443\u0441\u0442\u0438\u043c\u0430\u044f \u0432\u0435\u0440\u0441\u0438\u044f %s \u0434\u043b\u044f \u043e\u0431\u0440\u0430\u0442\u043d\u043e\u0433\u043e \u043f\u0440\u0435\u043e\u0431\u0440\u0430\u0437\u043e\u0432\u0430\u043d\u0438\u044f \u0438\u0437 \u043e\u0431\u044a\u0435\u043a\u0442\u0430 \u0431\u0435\u0437 \u0432\u0435\u0440\u0441\u0438\u0439\x00\u043e\u0440\u043a\u0435\u0441\u0442\u0440\u0430\u0442\u043e\u0440 \"%s\" \u043d\u0435 \u043f\u043e\u0434\u0434\u0435\u0440\u0436\u0438\u0432\u0430\u0435\u0442\u0441\u044f\x00\u0444\u0430\u0439\u043b \u0448\u0430\u0431\u043b\u043e\u043d\u0430 %s \u043d\u0435 \u0441\u0443\u0449\u0435\u0441\u0442\u0432\u0443\u0435\u0442\x00\u043d\u0435\u0440\u0430\u0441\u043f\u043e\u0437\u043d\u0430\u043d\u043d\u043e\u0435 \u0437\u043d\u0430\u0447\u0435\u043d\u0438\u0435 APIVersion \"%s\"\x00\u0444\u0430\u0439\u043b YAML %s \u043d\u0435 \u0441\u0443\u0449\u0435\u0441\u0442\u0432\u0443\u0435\u0442\x00")
+var _translationsRu_ruLc_messagesAcsengineMo = []byte("\xde\x12\x04\x95\x00\x00\x00\x00\x15\x00\x00\x00\x1c\x00\x00\x00\xc4\x00\x00\x00\x1d\x00\x00\x00l\x01\x00\x00\x00\x00\x00\x00\xe0\x01\x00\x00 \x00\x00\x00\xe1\x01\x00\x00+\x00\x00\x00\x02\x02\x00\x00F\x00\x00\x00.\x02\x00\x00I\x00\x00\x00u\x02\x00\x009\x00\x00\x00\xbf\x02\x00\x00=\x00\x00\x00\xf9\x02\x00\x00C\x00\x00\x007\x03\x00\x00!\x00\x00\x00{\x03\x00\x00(\x00\x00\x00\x9d\x03\x00\x00%\x00\x00\x00\xc6\x03\x00\x00\x19\x00\x00\x00\xec\x03\x00\x00\x19\x00\x00\x00\x06\x04\x00\x00.\x00\x00\x00 \x04\x00\x00*\x00\x00\x00O\x04\x00\x00+\x00\x00\x00z\x04\x00\x00>\x00\x00\x00\xa6\x04\x00\x00 \x00\x00\x00\xe5\x04\x00\x00\x1f\x00\x00\x00\x06\x05\x00\x00\x1c\x00\x00\x00&\x05\x00\x00\x1b\x00\x00\x00C\x05\x00\x00\x84\x01\x00\x00_\x05\x00\x00@\x00\x00\x00\xe4\x06\x00\x00?\x00\x00\x00%\a\x00\x00\x7f\x00\x00\x00e\a\x00\x00\x8c\x00\x00\x00\xe5\a\x00\x00y\x00\x00\x00r\b\x00\x00a\x00\x00\x00\xec\b\x00\x00^\x00\x00\x00N\t\x00\x00>\x00\x00\x00\xad\t\x00\x00P\x00\x00\x00\xec\t\x00\x00L\x00\x00\x00=\n\x00\x004\x00\x00\x00\x8a\n\x00\x002\x00\x00\x00\xbf\n\x00\x00Y\x00\x00\x00\xf2\n\x00\x00K\x00\x00\x00L\v\x00\x00\\\x00\x00\x00\x98\v\x00\x00\x87\x00\x00\x00\xf5\v\x00\x00=\x00\x00\x00}\f\x00\x004\x00\x00\x00\xbb\f\x00\x00=\x00\x00\x00\xf0\f\x00\x00*\x00\x00\x00.\r\x00\x00\x01\x00\x00\x00\f\x00\x00\x00\x02\x00\x00\x00\x04\x00\x00\x00\x06\x00\x00\x00\a\x00\x00\x00\t\x00\x00\x00\x0f\x00\x00\x00\x12\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\n\x00\x00\x00\x00\x00\x00\x00\x14\x00\x00\x00\x11\x00\x00\x00\b\x00\x00\x00\x13\x00\x00\x00\x0e\x00\x00\x00\x05\x00\x00\x00\x10\x00\x00\x00\x03\x00\x00\x00\x15\x00\x00\x00\x00\x00\x00\x00\v\x00\x00\x00\r\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00Error reading file %s, Error: %s\x00Error while querying ARM for resources: %+v\x00Found 2 resources with type %s in the template. There should only be 1\x00Found no resources with type %s in the template. There should have been 1\x00Total count of master VMs: %d exceeded expected count: %d\x00Upgrade to Kubernetes 1.6.2 is not supported from version: %s\x00Upgrade to Kubernetes version: %s is not supported from version: %s\x00error creating directory '%s': %s\x00error executing template for file %s: %v\x00error generating upgrade template: %s\x00error parsing file %s: %v\x00error reading file %s: %s\x00failed to create openssh public key string: %q\x00failed to generate private key for ssh: %q\x00failed to initialize template generator: %s\x00invalid version %s for conversion back from unversioned object\x00orchestrator '%s' is unsupported\x00template file %s does not exist\x00unrecognized APIVersion '%s'\x00yaml file %s does not exist\x00Project-Id-Version: acsengine\nReport-Msgid-Bugs-To: \nPOT-Creation-Date: 2017-07-25 00:04+0000\nPO-Revision-Date: 2017-07-24 17:23-0700\nLast-Translator: Jiangtian Li <jiangtianli@hotmail.com>\nLanguage-Team: English\nLanguage: en_US\nMIME-Version: 1.0\nContent-Type: text/plain; charset=UTF-8\nContent-Transfer-Encoding: 8bit\nPlural-Forms: nplurals=2; plural=(n != 1);\nX-Generator: Poedit 2.0.3\n\x00\u041e\u0448\u0438\u0431\u043a\u0430 \u043f\u0440\u0438 \u0447\u0442\u0435\u043d\u0438\u0438 \u0444\u0430\u0439\u043b\u0430 %s, \u043e\u0448\u0438\u0431\u043a\u0430: %s\x00\u041e\u0448\u0438\u0431\u043a\u0430 \u043f\u0440\u0438 \u0437\u0430\u043f\u0440\u043e\u0441\u0435 \u0440\u0435\u0441\u0443\u0440\u0441\u043e\u0432 \u0432 ARM: %+v\x00\u0412 \u0448\u0430\u0431\u043b\u043e\u043d\u0435 \u043d\u0430\u0439\u0434\u0435\u043d\u043e 2\u00a0\u0440\u0435\u0441\u0443\u0440\u0441\u0430 \u0441 \u0442\u0438\u043f\u043e\u043c %s. \u0422\u0430\u043a\u043e\u0439 \u0440\u0435\u0441\u0443\u0440\u0441 \u0434\u043e\u043b\u0436\u0435\u043d \u0431\u044b\u0442\u044c \u0432\u0441\u0435\u0433\u043e 1\x00\u0412 \u0448\u0430\u0431\u043b\u043e\u043d\u0435 \u043d\u0435 \u043d\u0430\u0439\u0434\u0435\u043d\u043e \u0440\u0435\u0441\u0443\u0440\u0441\u043e\u0432 \u0441 \u0442\u0438\u043f\u043e\u043c %s. \u0414\u043e\u043b\u0436\u0435\u043d \u043f\u0440\u0438\u0441\u0443\u0442\u0441\u0442\u0432\u043e\u0432\u0430\u0442\u044c 1 \u0442\u0430\u043a\u043e\u0439 \u0440\u0435\u0441\u0443\u0440\u0441\x00\u0427\u0438\u0441\u043b\u043e \u0433\u043b\u0430\u0432\u043d\u044b\u0445 \u0432\u0438\u0440\u0442\u0443\u0430\u043b\u044c\u043d\u044b\u0445 \u043c\u0430\u0448\u0438\u043d %d \u043f\u0440\u0435\u0432\u044b\u0441\u0438\u043b\u043e \u043e\u0436\u0438\u0434\u0430\u0435\u043c\u043e\u0435 \u0437\u043d\u0430\u0447\u0435\u043d\u0438\u0435: %d\x00\u041d\u0435 \u043f\u043e\u0434\u0434\u0435\u0440\u0436\u0438\u0432\u0430\u0435\u0442\u0441\u044f \u043e\u0431\u043d\u043e\u0432\u043b\u0435\u043d\u0438\u0435 \u0434\u043e Kubernetes\u00a01.6.2 \u0441 \u0432\u0435\u0440\u0441\u0438\u0438: %s\x00\u041d\u0435 \u043f\u043e\u0434\u0434\u0435\u0440\u0436\u0438\u0432\u0430\u0435\u0442\u0441\u044f \u043e\u0431\u043d\u043e\u0432\u043b\u0435\u043d\u0438\u0435 \u0434\u043e Kubernetes\u00a0%s \u0441 \u0432\u0435\u0440\u0441\u0438\u0438: %s\x00\u043e\u0448\u0438\u0431\u043a\u0430 \u043f\u0440\u0438 \u0441\u043e\u0437\u0434\u0430\u043d\u0438\u0438 \u043a\u0430\u0442\u0430\u043b\u043e\u0433\u0430 \"%s\": %s\x00\u043e\u0448\u0438\u0431\u043a\u0430 \u043f\u0440\u0438 \u0432\u044b\u043f\u043e\u043b\u043d\u0435\u043d\u0438\u0438 \u0448\u0430\u0431\u043b\u043e\u043d\u0430 \u0434\u043b\u044f \u0444\u0430\u0439\u043b\u0430 %s: %v\x00\u043e\u0448\u0438\u0431\u043a\u0430 \u043f\u0440\u0438 \u0441\u043e\u0437\u0434\u0430\u043d\u0438\u0438 \u0448\u0430\u0431\u043b\u043e\u043d\u0430 \u043e\u0431\u043d\u043e\u0432\u043b\u0435\u043d\u0438\u044f: %s\x00\u043e\u0448\u0438\u0431\u043a\u0430 \u043f\u0440\u0438 \u0430\u043d\u0430\u043b\u0438\u0437\u0435 \u0444\u0430\u0439\u043b\u0430 %s: %v\x00\u043e\u0448\u0438\u0431\u043a\u0430 \u043f\u0440\u0438 \u0447\u0442\u0435\u043d\u0438\u0438 \u0444\u0430\u0439\u043b\u0430 %s: %s\x00\u043d\u0435 \u0443\u0434\u0430\u043b\u043e\u0441\u044c \u0441\u043e\u0437\u0434\u0430\u0442\u044c \u0441\u0442\u0440\u043e\u043a\u0443 \u043e\u0442\u043a\u0440\u044b\u0442\u043e\u0433\u043e \u043a\u043b\u044e\u0447\u0430 OpenSSH: %q\x00\u043d\u0435 \u0443\u0434\u0430\u043b\u043e\u0441\u044c \u0441\u043e\u0437\u0434\u0430\u0442\u044c \u0437\u0430\u043a\u0440\u044b\u0442\u044b\u0439 \u043a\u043b\u044e\u0447 \u0434\u043b\u044f SSH: %q\x00\u043d\u0435 \u0443\u0434\u0430\u043b\u043e\u0441\u044c \u0438\u043d\u0438\u0446\u0438\u0430\u043b\u0438\u0437\u0438\u0440\u043e\u0432\u0430\u0442\u044c \u0433\u0435\u043d\u0435\u0440\u0430\u0442\u043e\u0440 \u0448\u0430\u0431\u043b\u043e\u043d\u043e\u0432: %s\x00\u043d\u0435\u0434\u043e\u043f\u0443\u0441\u0442\u0438\u043c\u0430\u044f \u0432\u0435\u0440\u0441\u0438\u044f %s \u0434\u043b\u044f \u043e\u0431\u0440\u0430\u0442\u043d\u043e\u0433\u043e \u043f\u0440\u0435\u043e\u0431\u0440\u0430\u0437\u043e\u0432\u0430\u043d\u0438\u044f \u0438\u0437 \u043e\u0431\u044a\u0435\u043a\u0442\u0430 \u0431\u0435\u0437 \u0432\u0435\u0440\u0441\u0438\u0439\x00\u043e\u0440\u043a\u0435\u0441\u0442\u0440\u0430\u0442\u043e\u0440 \"%s\" \u043d\u0435 \u043f\u043e\u0434\u0434\u0435\u0440\u0436\u0438\u0432\u0430\u0435\u0442\u0441\u044f\x00\u0444\u0430\u0439\u043b \u0448\u0430\u0431\u043b\u043e\u043d\u0430 %s \u043d\u0435 \u0441\u0443\u0449\u0435\u0441\u0442\u0432\u0443\u0435\u0442\x00\u043d\u0435\u0440\u0430\u0441\u043f\u043e\u0437\u043d\u0430\u043d\u043d\u043e\u0435 \u0437\u043d\u0430\u0447\u0435\u043d\u0438\u0435 APIVersion \"%s\"\x00\u0444\u0430\u0439\u043b YAML %s \u043d\u0435 \u0441\u0443\u0449\u0435\u0441\u0442\u0432\u0443\u0435\u0442\x00")
 
 func translationsRu_ruLc_messagesAcsengineMoBytes() ([]byte, error) {
 	return _translationsRu_ruLc_messagesAcsengineMo, nil
@@ -3594,11 +3594,13 @@ var _bindata = map[string]func() (*asset, error){
 // directory embedded in the file by go-bindata.
 // For example if you run go-bindata on data/... and data contains the
 // following hierarchy:
-//     data/
-//       foo.txt
-//       img/
-//         a.png
-//         b.png
+//
+//	data/
+//	  foo.txt
+//	  img/
+//	    a.png
+//	    b.png
+//
 // then AssetDir("data") would return []string{"foo.txt", "img"}
 // AssetDir("data/img") would return []string{"a.png", "b.png"}
 // AssetDir("foo.txt") and AssetDir("notexist") would return an error
